backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
celery[redis]==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2
django-extensions==3.2.3

# Testing
pytest==8.3.4
pytest-django==4.9.0
